Refactor code to separate rental management logic into its own class
