+++ app.rb
@@ -2,6 +2,7 @@ require_relative 'book'
 require_relative 'rental'
 require_relative 'student'
 require_relative 'teacher'
+require_relative 'rental_manager'
 
 # Class responsible for book management
 class BookManager
@@ -99,16 +100,17 @@ class App
     puts 'Welcome to School Library App!'
     @book_manager = BookManager.new
     @person_manager = PersonManager.new
+    @rental_manager = RentalManager.new(@book_manager, @person_manager)
   end
-
+  
   def handle_user_choice(choice)
     actions = {
       '1' => -> { @book_manager.list_all_books },
       '2' => -> { @person_manager.list_all_people },
       '3' => -> { @person_manager.create_person },
       '4' => -> { @book_manager.create_book },
-      '5' => -> { create_new_rental },
-      '6' => -> { list_all_rentals_for_person },
+      '5' => -> { @rental_manager.create_new_rental },
+      '6' => -> { @rental_manager.list_all_rentals_for_person },
       '7' => lambda {
                puts 'Thank you for using this app!'
                exit
@@ -119,66 +121,6 @@ class App
       actions[choice].call
     else
       puts 'Invalid option. Please try again.'
-    end
-  end
-
-  def select_book(books)
-    @book_manager.list_all_books
-    book_index = gets.chomp.to_i
-    return nil unless book_index.between?(0, books.length - 1)
-
-    book_index
-  end
-
-  def select_person(people)
-    @person_manager.list_all_people
-    person_index = gets.chomp.to_i
-    return nil unless person_index.between?(0, people.length - 1)
-
-    person_index
-  end
-
-  def create_new_rental
-    return unless valid_conditions_for_rental?
-
-    book_index = select_book(@book_manager.books)
-    return if book_index.nil?
-
-    person_index = select_person(@person_manager.people)
-    return if person_index.nil?
-
-    print "\nDate: "
-    date = gets.chomp
-    new_rental = Rental.new(date, @book_manager.books[book_index], @person_manager.people[person_index])
-    @book_manager.books[book_index].add_rental(new_rental)
-    puts 'Rental created successfully'
-  end
-
-  def valid_conditions_for_rental?
-    if @book_manager.books.empty? || @person_manager.people.empty?
-      puts 'There must be at least one book and one person to create a rental.'
-      false
-    else
-      true
-    end
-  end
-
-  def list_all_rentals_for_person
-    print 'ID of person: '
-    id = gets.chomp.to_i
-    person = @person_manager.people.find { |p| p.id == id }
-    if person.nil?
-      puts 'Person not found.'
-      return
-    end
-    rentals = person.rentals
-    if rentals.empty?
-      puts "No rentals found for person with ID #{id}."
-    else
-      puts 'Rentals:'
-      rentals.each do |rental|
-        puts "Date: #{rental.date}, Book \"#{rental.book.title}\" by #{rental.book.author}"
-      end
     end
   end
 end
